+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,13 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".SkillActivity">
+        <activity android:name=".Controller.SkillActivity">
         </activity>
-        <activity android:name=".BaseActivity">
+        <activity android:name=".Controller.BaseActivity">
         </activity>
-        <activity android:name=".LeagueActivity">
+        <activity android:name=".Controller.LeagueActivity">
         </activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".Controller.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
