--- app/src/main/java/com/ehword/swoosh/LeagueActivity.kt
@@ -1,23 +1,0 @@
-package com.ehword.swoosh
-
-import android.content.Intent
-import android.view.View
-import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_league.*
-
-class LeagueActivity : BaseActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_league)
-
-//        leagueNextButton.setOnClickListener {
-//            val skillIntent = Intent(this, SkillActivity::class.java)
-//            startActivity(skillIntent)
-//        }
-    }
-    fun leagueNextClicked (view: View){
-        val skillIntent = Intent(this, SkillActivity::class.java)
-        startActivity(skillIntent)
-    }
-}
