+++ app/src/main/java/com/ehword/swoosh/Utilities/Constants.kt
@@ -1,0 +1,3 @@
+package com.ehword.swoosh.Utilities
+
+const val EXTRA_LEAGUE = "league"
