+++ app/src/main/res/layout/activity_league.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".LeagueActivity">
+        tools:context=".Controller.LeagueActivity">
 
     <ImageView
             android:layout_width="0dp"
@@ -52,19 +52,22 @@
                 android:background="@drawable/swoosh_toogle" android:fontFamily="@font/montserrat"
                 android:textColor="@android:color/white" android:layout_marginTop="16dp"
                 android:layout_marginBottom="16dp" android:textOff="@string/mens" android:textOn="@string/mens"
-                android:textStyle="bold" android:textSize="18sp"/>
+                android:textStyle="bold" android:textSize="18sp" android:onClick="mensButtonClicked"/>
         <ToggleButton
                 android:text="@string/womens"
                 android:layout_width="match_parent"
-                android:layout_height="58dp" android:id="@+id/toggleButton2" android:textOff="@string/womens"
-                android:textOn="@string/womens" android:textColor="@android:color/white" android:layout_marginBottom="16dp"
-                android:background="@drawable/swoosh_toogle" android:textSize="18sp" android:textStyle="bold"/>
+                android:layout_height="58dp" android:id="@+id/womensButton" android:textOff="@string/womens"
+                android:textOn="@string/womens" android:textColor="@android:color/white"
+                android:layout_marginBottom="16dp"
+                android:background="@drawable/swoosh_toogle" android:textSize="18sp" android:textStyle="bold"
+                android:onClick="womensButtonClicked"/>
         <ToggleButton
                 android:text="@string/coed"
                 android:layout_width="match_parent"
-                android:layout_height="58dp" android:id="@+id/toggleButton3" android:textOff="@string/coed"
+                android:layout_height="58dp" android:id="@+id/coedButton" android:textOff="@string/coed"
                 android:textOn="@string/coed" android:textColor="@android:color/white"
-                android:background="@drawable/swoosh_toogle" android:textSize="18sp" android:textStyle="bold"/>
+                android:background="@drawable/swoosh_toogle" android:textSize="18sp" android:textStyle="bold"
+                android:onClick="coedButtonClicked"/>
     </LinearLayout>
     <Button
             android:text="@string/next"
