+++ app/src/main/res/values/strings.xml
@@ -15,5 +15,4 @@
     <string name="baller">Baller</string>
     <string name="finish">Finish</string>
     <string name="i_am_a">I am a</string>
-    <string name="background_image">Background Image</string>
 </resources>
